This commit was manufactured by cvs2git to create tag 'CondTools- Ecal-V00-14-01-04'.
Sprout from master 2009-10-27 15:44:57 UTC Emilia Lubenova Becheva <sha1-d966cd138a0288d3ff259f3f2cdb9f6c54000612@cern.ch> 'bug fix for warning messages during the compilation'
Cherrypick from master 2010-03-04 16:41:21 UTC Emanuele Di Marco <sha1-7725c4b6a787836a48e2e74f4d20b6a79f483406@cern.ch> 'adding two errors for LED amplitude':
    interface/EcalChannelStatusHandler.h
    interface/EcalErrorDictionary.h
    interface/EcalErrorMask.h
    src/EcalErrorMask.cc
Cherrypick from master 2009-11-03 16:12:56 UTC Francesca Cavallari <sha1-cf24814f6426e55741f66a9979d2192918cf784d@cern.ch> 'p1 p2 version for 33X P5':
    src/EcalLaserAPDPNRatiosXMLTranslator.cc
    test/testXMLTranslators.cc
Cherrypick from master 2009-05-14 08:26:22 UTC Francesca Cavallari <sha1-cf24814f6426e55741f66a9979d2192918cf784d@cern.ch> 'improved handling of iovs and check on APD/PN':
    src/EcalLaserHandler.cc

+++ interface/EcalChannelStatusHandler.h
@@ -28,7 +28,7 @@
 #include "FWCore/Framework/interface/EventSetup.h"
 #include "FWCore/Framework/interface/EventSetupRecordKey.h"
 
-#include "DQM/EcalCommon/interface/EcalErrorMask.h"
+#include "CondTools/Ecal/interface/EcalErrorMask.h"
 #include "CondTools/Ecal/interface/EcalErrorDictionary.h"
 
 #include "CondFormats/EcalObjects/interface/EcalChannelStatus.h"
@@ -112,8 +112,8 @@ namespace popcon {
       
     private:
       
-      unsigned long m_firstRun ;
-      unsigned long m_lastRun ;
+      unsigned int m_firstRun ;
+      unsigned int m_lastRun ;
       
       std::string m_location;
       std::string m_gentag;
